requirements.txt
Flask~=2.1.1
marshmallow~=3.15.0
Flask_Cors
Flask-Caching
Flask-RESTful
PyYaml
